chore : 실제 서비스용 (app) 라우트 그룹 추가
