+++ frontend/src/app/(app)/(protected)/history/layout.tsx
