+++ frontend/src/app/(app)/(protected)/history/page.tsx
