+++ frontend/src/app/(app)/(protected)/layout.tsx
