+++ frontend/src/app/(app)/(protected)/reservation/complete/[waitingId]/page.tsx
