+++ frontend/src/app/(app)/(protected)/reservation/detail/[waitingId]/page.tsx
