+++ frontend/src/app/(app)/(protected)/waiting/[waitingId]/page.tsx
