+++ frontend/src/app/(app)/(with-navbar)/(home)/layout.tsx
