+++ frontend/src/app/(app)/(with-navbar)/layout.tsx
