+++ frontend/src/app/(app)/(with-navbar)/map/page.tsx
