+++ frontend/src/app/(app)/detail/[popupId]/data.ts
