+++ frontend/src/app/(app)/detail/[popupId]/map/page.tsx
