+++ frontend/src/app/(app)/detail/[popupId]/page.tsx
