+++ frontend/src/app/landing/layout.tsx
@@ -1,0 +1,9 @@
+export default function LandingLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="min-h-screen w-full bg-black text-white">{children}</div>
+  );
+}
